fix(python-core): accept float bezier coordinates in Shape.from_string
sompyler-web instruments use float x-coordinates (0-1 normalized range)
and float y-values (sub-integer precision) in shape strings, e.g.:
  A: "2.0:1;0.05,40;0.15,70;0.5,90;1,100"
  R: "0.001:100;0.05,50;0.3,0.5;0.5,0.05;1,0"

The Python parser used int(x) / int(y) which raises ValueError on these.
Changed to float(x) / float(y); the normalization in Shape.__init__ already
divides by x_max and y_max, so float/int coordinates produce identical
normalized output — the change is backward-compatible.

+++ public/python-core/Sompyler/synthesizer/shape/__init__.py
@@ -218,7 +218,7 @@ class Shape:
                     else:
                         z = 1
                     syntracer("coords", x=x, y=y, z=z, is_sharp=s)
-                    coords.append([ int(x), int(y), int(z), s ])
+                    coords.append([ float(x), float(y), int(z), s ])
             except (ValueError,TypeError) as e:
                 raise ShapeDefinitionError(f"{m0}: illegal syntax") from e
